requirements.txt
brotlipy
aspose-words
pdf2pptx
tabula-py
pdf2docx
boto3
Flask